Moved command_line content to cli, pyreportcard-cli to cli/main
